Before Start Again
At this point I've realized the extreme importance of mobile-firts development, because it's too difficult to transform the design I did into smaller screens.

So I'm deciding to start the benzá project again, this time taking a mobile-first approach. I believe the code and the project in general will be better.